[refactor] Extracted test for pytest.skip into a separate test module.
This prevents unrelated tests from aggregating in test_simple.py.
