+++ tests/test_simple.py
@@ -260,41 +260,6 @@ class TestEventLoopStartedBeforeFixtures:
         assert await loop.run_in_executor(None, self.foo) == 1
 
 
-def test_asyncio_marker_compatibility_with_skip(pytester: Pytester):
-    pytester.makepyfile(
-        dedent(
-            """\
-                import pytest
-
-                pytest_plugins = "pytest_asyncio"
-
-                @pytest.mark.asyncio
-                async def test_no_warning_on_skip():
-                    pytest.skip("Test a skip error inside asyncio")
-            """
-        )
-    )
-    result = pytester.runpytest("--asyncio-mode=strict")
-    result.assert_outcomes(skipped=1)
-
-
-def test_asyncio_auto_mode_compatibility_with_skip(pytester: Pytester):
-    pytester.makepyfile(
-        dedent(
-            """\
-                import pytest
-
-                pytest_plugins = "pytest_asyncio"
-
-                async def test_no_warning_on_skip():
-                    pytest.skip("Test a skip error inside asyncio")
-            """
-        )
-    )
-    result = pytester.runpytest("--asyncio-mode=auto")
-    result.assert_outcomes(skipped=1)
-
-
 def test_invalid_asyncio_mode(testdir):
     result = testdir.runpytest("-o", "asyncio_mode=True")
     result.stderr.no_fnmatch_line("INTERNALERROR> *")
